feat: move init to app
